Move call presentation from the chats flow into the user session flow. (#4459)
