+++ ElementX/Sources/Screens/UserProfileScreen/UserProfileScreenCoordinator.swift
@@ -18,7 +18,7 @@ struct UserProfileScreenCoordinatorParameters {
 
 enum UserProfileScreenCoordinatorAction {
     case openDirectChat(roomID: String)
-    case startCall(roomID: String)
+    case startCall(roomProxy: JoinedRoomProxyProtocol)
     case dismiss
 }
 
@@ -47,8 +47,8 @@ final class UserProfileScreenCoordinator: CoordinatorProtocol {
             switch action {
             case .openDirectChat(let roomID):
                 actionsSubject.send(.openDirectChat(roomID: roomID))
-            case .startCall(let roomID):
-                actionsSubject.send(.startCall(roomID: roomID))
+            case .startCall(let roomProxy):
+                actionsSubject.send(.startCall(roomProxy: roomProxy))
             case .dismiss:
                 actionsSubject.send(.dismiss)
             }
