+++ ElementX/Sources/Screens/UserProfileScreen/UserProfileScreenViewModel.swift
@@ -62,7 +62,7 @@ class UserProfileScreenViewModel: UserProfileScreenViewModelType, UserProfileScr
         case .createDirectChat:
             Task { await createDirectChat() }
         case .startCall(let roomID):
-            actionsSubject.send(.startCall(roomID: roomID))
+            Task { await startCall(roomID: roomID) }
         case .dismiss:
             actionsSubject.send(.dismiss)
         }
@@ -143,12 +143,21 @@ class UserProfileScreenViewModel: UserProfileScreenViewModelType, UserProfileScr
         }
     }
     
-    // MARK: Loading indicator
+    private func startCall(roomID: String) async {
+        guard case let .joined(roomProxy) = await userSession.clientProxy.roomForIdentifier(roomID) else {
+            showErrorIndicator()
+            return
+        }
+        actionsSubject.send(.startCall(roomProxy: roomProxy))
+    }
     
-    private static let loadingIndicatorIdentifier = "\(UserProfileScreenViewModel.self)-Loading"
+    // MARK: User Indicators
+    
+    private var loadingIndicatorIdentifier: String { "\(Self.self)-Loading" }
+    private var statusIndicatorIdentifier: String { "\(Self.self)-Status" }
     
     private func showLoadingIndicator(allowsInteraction: Bool) {
-        userIndicatorController.submitIndicator(UserIndicator(id: Self.loadingIndicatorIdentifier,
+        userIndicatorController.submitIndicator(UserIndicator(id: loadingIndicatorIdentifier,
                                                               type: .modal(progress: .indeterminate, interactiveDismissDisabled: false, allowsInteraction: allowsInteraction),
                                                               title: L10n.commonLoading,
                                                               persistent: true),
@@ -156,6 +165,13 @@ class UserProfileScreenViewModel: UserProfileScreenViewModelType, UserProfileScr
     }
     
     private func hideLoadingIndicator() {
-        userIndicatorController.retractIndicatorWithId(Self.loadingIndicatorIdentifier)
+        userIndicatorController.retractIndicatorWithId(loadingIndicatorIdentifier)
+    }
+    
+    private func showErrorIndicator() {
+        userIndicatorController.submitIndicator(UserIndicator(id: statusIndicatorIdentifier,
+                                                              type: .toast,
+                                                              title: L10n.errorUnknown,
+                                                              iconName: "xmark"))
     }
 }
